Make action usable for github actions
GitHub actions expects the checked-in action to be usable "as is"
without running any additional commands (like npm install). Thus we need
to check in a "bundled" version. But since GH actions can't do the
bundling for us (duh!), we need to manually do it and have an action
that makes sure we don't forget...